Refactor importsummary remote data source(not a domain concept) as a base class
